DOCS-0000: Update onboarding reference in getting started
Recently, the onboarding flow has changed.
Previously, account holders were asked to supply information
through forms and mail. Currently, this information is supplied
through a series of webforms in MultiSafepay Control. This is
also called 'onboarding flow' or 'task-based onboarding'.

This commit updates the getting started page to correctly refer
to the process of supplying information through this new method.